A function to provide a list of Wireshark profiles
The command tshark -G folders is used to determine the locations - both
global and personal. I will use this to allow the user to connect a
tshark profile to a Wireshark profile, so that when tshark is invoked,
it can be influenced by a specific Wireshark profile (e.g. for packet
colors)