Rework profile minibuffer commands and add a connection to Wireshark
This change reworks profile support. First, when the create-profile UI
is opened, a new section at the bottom of the dialog allows making a
connection to a Wireshark profile. If the checkbox is ticked, then the
drop down button can be used to choose a Wireshark profile. This is
saved in termshark.toml (for the current termshark profile) as
main.wireshark-profile (string). The effect is that when tshark is
invoked e.g. to load the PSML or PDML for a pcap, the -C <profile> flag
is provided. One practical effect of this is that the Wireshark packet
colors the user may have set up under a specific profile are able to be
used in termshark. So, this mechanism allows configuration of termshark
profiles via a connection to a Wireshark profile. Note that most
Wireshark settings are not adopted in this way; for example, column sets
can be configured in Wireshark, but termshark does this independently
and stores them in the user's current termshark profile. You can change
column sets by changing termshark profile, but they are stored by
termshark in its config and not consumed from Wireshark's config.

The second change reworks the profile minibuffer commands. Instead of
profile and new-profile, there is a multi-level profile command:

- profile
  - create
  - use <profile>
  - delete <profile>
  - link <wireshark profile>
  - unlink