Fix capture method comparison
If the default is "termshark", then termshark's dual-pronged
dumpcap-then-tshark capture method is used. But it's misleading to print
out that when termshark's capture failed, here is the stderr from the
failing capture program, "termshark". So I swap it out for some other
language in the default case. This compares the binaries properly, via
their full paths - if termshark is started as e.g. ./termshark, my
comparison failed.