Fix a race condition that sporadically breaks extcap captures
From the code comments:

To capture packets, termshark runs itself with a special env var set.
It detects this at startup, then launches dumpcap as the first capture
method. If this fails (e.g. the source is an extcap), it launches tshark
instead. Dumpcap is more efficient, but tshark is needed for the extcap
sources. The problem is that termshark needs a heuristic for when
packets have actually been detected - this is so it can wait to launch
the UI (in case a password is needed at the terminal, I don't want to
obscure that with the UI). So termshark waits for a WRITE to the pcap
generated by the capture process, and then launches tail. BUT - if
dumpcap fails, it will delete (unlink) the capture file passed to it
with the -w argument before tshark starts. If we don't watch for WRITE,
this triggers the notifier; then tail starts; then tail fails because
depending on timing, the fallback tshark -i process may not have started
yet and so the tail target pcap does not exist.  The fix is to monitor
for inotify REMOVE too, and if seen, recreate the pcap file (empty), and
restart the watcher. And importantly, don't let the tail process start
until the WRITE event is seen.