Add a summary of stderr to the dialog displayed when tshark fails
If a tshark process - run by termshark - returns a non-zero exit code,
termshark will send that code and an error string to a UI dialog. But
often this is not enough to help you diagnose the problem, and you have
to dig through termshark's log file for more info. This change adds a
summary of stderr to the error dialog displayed, specifically the first
two and last two lines read from the stderr of the failed tshark
process.