Bug fix to prevent packet list search not returning
The symptom is that the spinner stays active. When the packet list
search wraps, it mistakenly moves back to packet 1 to resume the
search. But if a display filter is applied, packet 1 may not be
loaded. The search implementation will make a request to the loader to
load packet 1, but the loader ignores it because as far as it's
concerned, the first block of packets have been loaded in their
entirety; they just don't include packet 1 because it doesn't match the
display filter. This procedure repeated indefinitely. The fix is to loop
back to the first packet loaded.