Change default suppress-errors boolean to true, not false
If this value is not set explicitly, a false value means that an error
from a termshark-initiated tshark process will result in an error dialog
in the termshark UI. I am seeing this more and more as I test with
various pcaps - it always comes from tshark serializing characters into
XML text that are invalid, according to the XML spec (val <= 31 and val
not in {tab, CR, LF}). Here is a merge request against Wireshark to try
to solve this problem at the source:
https://gitlab.com/wireshark/wireshark/-/merge_requests/7398

To see the problem, try this:

$ wget https://storage.googleapis.com/gcla3/xmlbug.pcapng
$ tshark -r xmlbug.pcapng -T pdml | xmllint --noout - || echo bad xml

Even if this Wireshark request is merged, it will presumably be a long
time before all termshark-used tsharks are updated. So I think the more
user-friendly option is to suppress these errors to avoid popups about
which the user can do very little anyway.

Here's a hack you can use if you want to see errors, in general, but are
not interested in this specific XML error:

https://github.com/gcla/termshark/issues/133#issuecomment-985194719

Workaround:

- create the following file called e.g. /usr/local/bin/tshark-hack

if [[ " $* " =~ " pdml " ]]; then
    exec tshark "$@" | tr -cd '\11\12\15\40-\176'
else
    exec tshark "$@"
fi

- run:

$ sudo chmod +x /usr/local/bin/tshark-hack

- edit ~/.config/termshark/termshark.toml

[main]
  tshark = "/usr/local/bin/tshark-hack"