Fix bug meaning that default config was not reloaded
When I refactored the config code to support profiles, I broke this
functionality - the correct call restores the ability of termshark to
respond to config file edits that happen outside of the running
application. Note that to fully support this, I should add a watcher for
the profile config too (coming soon).